Stops treating empty spaces as sprites
This was dumb, I'm not sure why I did that in the first place. In
theory, this should also give us a speed boost on large maps, since the
blank spaces don't need to be iterated through each frame. Probably not
that large a speed boost though, since it's not like they were being
rendered.